requirements.txt
flask
torch
openai-whisper
gtts
ffmpeg-python
